+++ iqm-qaoa/src/iqm/applications/maxcut.py
@@ -47,16 +47,17 @@ Example:
 
 from collections.abc import Iterator
 import random
+from typing import Literal
 
 import cvxpy as cp
 from dimod import BinaryQuadraticModel
+from iqm.applications.graph_utils import _generate_desired_graph
 from iqm.applications.qubo import QUBOInstance, relabel_graph_nodes
 import networkx as nx
 import numpy as np
 from scipy.linalg import eigh
 
 
-# pylint: disable=anomalous-backslash-in-string
 class MaxCutInstance(QUBOInstance):
     r"""The maxcut instance class.
 
@@ -140,7 +141,6 @@ class MaxCutInstance(QUBOInstance):
         return cut
 
 
-# pylint: disable=anomalous-backslash-in-string
 class WeightedMaxCutInstance(QUBOInstance):
     r"""The weighted maxcut instance class.
 
@@ -232,17 +232,17 @@ class WeightedMaxCutInstance(QUBOInstance):
         return cut
 
 
-# pylint: disable=anomalous-backslash-in-string
-def maxcut_generator(
+def maxcut_generator(  # noqa: PLR0913
     n: int,
     n_instances: int,
     *,
-    graph_family: str = "erdos-renyi",
+    graph_family: Literal["regular", "erdos-renyi"] = "erdos-renyi",
     p: float = 0.5,
     d: int = 3,
     break_z2: bool = False,
-    seed: int = 1337,
+    seed: int | None = None,
     enforce_connected: bool = False,
+    max_iterations: int = 1000,
 ) -> Iterator[MaxCutInstance]:
     r"""The generator function for generating random maxcut problem instances.
 
@@ -262,27 +262,18 @@ def maxcut_generator(
         break_z2: Optional bool indicating whether the :math:`\mathbb{Z}_2` symmetry should be explicitly broken
             in the problem instances.
         seed: Optional random seed for generating the problem instances.
-        enforce_connected: A bool stating whether it is required that the random graphs are connected.
+        enforce_connected: True iff it is required that the random graphs are connected.
+        max_iterations: In case ``enforce_connected`` is ``True``, the function generates random graphs in a ``while``
+            loop until it finds a connected one. If it doesn't find a connected one after ``max_iterations``, it raises
+            an error.
 
     Yields:
         Problem instances of :class:`MaxCutInstance` randomly constructed in accordance to the input parameters.
 
     """
-
-    def _generate_desired_graph(graph_family: str, n: int, p: float, d: int, seed: int) -> nx.Graph:
-        """Wrapper helper function to encapsulate the graph generation logic."""
-        if graph_family == "erdos-renyi":
-            return nx.erdos_renyi_graph(n, p, seed)
-        if graph_family == "regular":
-            return nx.random_regular_graph(d, n, seed)
-        raise ValueError("Invalid random graph type. Choose either 'regular' or 'erdos-renyi'.")
-
     for _ in range(n_instances):
-        while True:
-            g = _generate_desired_graph(graph_family, n, p, d, seed)
-            seed = seed + 1  # Increment ``seed``, so that we don't keep generating the same graph over and over again.
-            if nx.is_connected(g) or not enforce_connected:
-                break
+        g = _generate_desired_graph(graph_family, n, p, d, seed, enforce_connected, max_iterations)
+
         yield MaxCutInstance(g, break_z2=break_z2)
 
 
@@ -328,7 +319,6 @@ def greedy_max_cut(max_cut_problem: MaxCutInstance | nx.Graph) -> str:
     return "".join(current_solution)
 
 
-# pylint: disable=too-many-locals
 def goemans_williamson(max_cut_problem: MaxCutInstance | nx.Graph) -> str:
     """Runs the Goemans-Williamson algorithm for maxcut, returning a solution bitstring.